protection of actions with user products

+++ server/routes/product.routes.js
@@ -1,6 +1,8 @@
 const express = require('express')
 const Product = require('../models/Product')
+const User = require('../models/User')
 const router = express.Router({ mergeParams: true })
+const auth = require('../middleware/auth.middleware')
 
 router
     .route('/')
@@ -14,13 +16,20 @@ router
             })
         }
     })
-    .post(async (req, res) => {
-        try {
-            const newProduct = await Product.create({ ...req.body })
-            res.status(201).send(newProduct)
-        } catch (e) {
-            res.status(500).json({
-                message: `На сервере произошла ошибка ${e.message}`
+    .post(auth, async (req, res) => {
+        const { admin } = await User.findOne({ _id: req.user._id })
+        if (admin) {
+            try {
+                const newProduct = await Product.create({ ...req.body })
+                res.status(201).send(newProduct)
+            } catch (e) {
+                res.status(500).json({
+                    message: `На сервере произошла ошибка ${e.message}`
+                })
+            }
+        } else {
+            res.status(401).json({
+                message: 'У вас нет доступа'
             })
         }
     })
@@ -39,32 +48,48 @@ router
             })
         }
     })
-    .delete(async (req, res) => {
-        try {
-            const { productId } = req.params
-            if (productId) {
-                await Product.findByIdAndDelete(productId)
-                res.send(productId)
-            }
+    .delete(auth, async (req, res) => {
+        const { admin } = await User.findOne({ _id: req.user._id })
+        if (admin) {
+            try {
+                const { productId } = req.params
+                if (productId) {
+                    await Product.findByIdAndDelete(productId)
+                    res.send(productId)
+                }
 
-        } catch (e) {
-            res.status(500).json({
-                message: `На сервере произошла ошибка: ${e}`
-            })
-        }
-    })
-    .patch(async (req, res) => {
-        try {
-            const {productId} = req.params
-            if (productId) {
-                const updateProduct = await Product.findByIdAndUpdate(productId, req.body, {new: true})
-                res.send(updateProduct)
+            } catch (e) {
+                res.status(500).json({
+                    message: `На сервере произошла ошибка: ${e}`
+                })
             }
-        } catch (e) {
-            res.status(500).json({
-                message: `На сервере произошла ошибка: ${e}`
+        } else {
+            res.status(401).json({
+                message: 'У вас нет доступа'
             })
         }
+
+    })
+    .patch(auth, async (req, res) => {
+        const { admin } = await User.findOne({ _id: req.user._id })
+        if (admin) {
+            try {
+                const { productId } = req.params
+                if (productId) {
+                    const updateProduct = await Product.findByIdAndUpdate(productId, req.body, { new: true })
+                    res.send(updateProduct)
+                }
+            } catch (e) {
+                res.status(500).json({
+                    message: `На сервере произошла ошибка: ${e}`
+                })
+            }
+        } else {
+            res.status(401).json({
+                message: 'У вас нет доступа'
+            })
+        }
+
     })
 
 module.exports = router
